[ADD] 요청보내는 순간의 날짜와 시간 index값 계산해주는 함수 추가
