+++ src/pages/ReserveHome.jsx
@@ -10,19 +10,17 @@ import { searchDayTable } from '../api/reservation';
 import ConfirmInfo from '../components/reservation/ConfirmInfo';
 import ReserveConfirm from '../components/reservation/ReserveConfirm';
 import { setPage } from '../App';
-import { timeToLange } from '../utils/timeToIndex';
+import { requestDate, timeToLange } from '../utils/requestDateInfo';
 export const reserveConfirm = createContext();
 export const fillReserveInfo = createContext();
 export const pickDate = createContext();
 
 function ReserveHome() {
+  const today = new Date();
   // getTimezoneOffset()은 현재 시간과의 차이를 분 단위로 반환하는 함수 => 9시간 이므로 -540 이 나올거임
   const offset = new Date().getTimezoneOffset() * 60000; // ms 단위를 맞추기 위해 60000 곱해줌
-  // const today = new Date(Date.now() - offset);
-  const today = new Date();
   const offsetToday = new Date(Date.now() - offset);
   const nowDate = offsetToday.toISOString().split('T')[0];
-  // console.log(nowDate);
 
   const { componentPage, setComponentPage } = useContext(setPage);
   const [resInfo, setResInfo] = useState({
@@ -31,11 +29,11 @@ function ReserveHome() {
     date: null,
     start: null,
     end: null,
-    people_num: null,
-    current_date: nowDate,
+    people_num: 2,
+    current_date: requestDate(),
     current_index: timeToLange(),
   });
-  console.log('test: ', resInfo);
+  console.log('예약정보 실시간 파악 : ', resInfo);
 
   useEffect(() => {
     /** suggest 페이지 초기 로딩 시, componentPage 숫자 1로 초기화
